+++ packages/inferno/src/DOM/patching.ts
@@ -4,15 +4,14 @@ import { directClone } from '../core/implementation';
 import { VNode } from '../core/types';
 import { mount, mountArrayChildren, mountTextContent, mountWasabyCallback, getMarkupForTemplatedNode } from './mounting';
 import { clearDOM, remove, removeAllChildren, unmount, unmountAllChildren } from './unmounting';
-import { appendChild, createDerivedState, EMPTY_OBJ, findDOMfromVNode, moveVNodeDOM, options, removeChild, removeVNodeDOM, replaceChild, callAll } from './utils/common';
+import { appendChild, createDerivedState, EMPTY_OBJ, findDOMfromVNode, moveVNodeDOM, options, removeChild, removeVNodeDOM, replaceChild } from './utils/common';
 import { isControlledFormElement, processElement } from './wrappers/processElement';
 import { patchProp } from './props';
 import { handleComponentInput, renderNewInput } from './utils/componentutil';
 import { validateKeys } from '../core/validate';
 import { mountRef, unmountRef } from '../core/refs';
-import { getDecoratedMarkup, nextTickWasaby } from '../wasaby/control'
+import { getDecoratedMarkup } from '../wasaby/control'
 
-export const QUEUE = [];
 
 function replaceWithNewNode(lastVNode, nextVNode, parentDOM: Element, context: Object, isSVG: boolean, lifecycle: Function[], environment?: any, parentControlNode?: any) {
   unmount(lastVNode);
@@ -364,56 +363,6 @@ function patchChildren(
   }
 }
 
-function updateWasabyControl(controlNode, parentDOM, lifecycle) {
-  let shouldUp;
-  try {
-      let resolvedContext;
-      //  Logger.log('DirtyChecking (update node with changed)', [
-      //      '',
-      //      '',
-      //      changedOptions || changedInternalOptions || changedAttrs || changedContext
-      //  ]);
-      controlNode.environment.setRebuildIgnoreId(controlNode.id);
-      // @ts-ignore
-      OptionsResolver.resolveInheritOptions(controlNode.controlClass, controlNode, controlNode.options);
-      controlNode.control.saveInheritOptions(controlNode.inheritOptions);
-      // @ts-ignore
-      resolvedContext = ContextResolver.resolveContext(controlNode.controlClass, controlNode.context, controlNode.control);
-      // Forbid force update in the time between _beforeUpdate and _afterUpdate
-      // @ts-ignore
-      const beforeUpdateResults = controlNode.control._beforeUpdate && controlNode.control.__beforeUpdate(controlNode.control._options, resolvedContext);
-      // controlNode.control._options = newOptions;
-      // @ts-ignore
-      const shouldUpdate = (controlNode.control._shouldUpdate ? controlNode.control._shouldUpdate(controlNode.control._options, resolvedContext) : true);
-      // controlNode.control._setInternalOptions(changedInternalOptions || {});
-      controlNode.oldOptions = controlNode.options; // TODO Для afterUpdate подумать, как еще можно передать
-      // TODO Для afterUpdate подумать, как еще можно передать
-      controlNode.oldContext = controlNode.context; // TODO Для afterUpdate подумать, как еще можно передать
-      // TODO Для afterUpdate подумать, как еще можно передать
-      // controlNode.attributes = nextVNode.controlAttributes;
-      // controlNode.events = nextVNode.controlEvents;
-      controlNode.control._saveContextObject(resolvedContext);
-      // @ts-ignore
-      controlNode.control.saveFullContext(ContextResolver.wrapContext(controlNode.control, controlNode.control._context));
-  }
-  finally {
-      /**
-       * TODO: удалить после синхронизации с контролами
-       */
-      shouldUp = controlNode.control._shouldUpdate ? controlNode.control._shouldUpdate(controlNode.control._options, controlNode.context) : true;
-  }
-  if (shouldUp) {
-      // @ts-ignore
-      const nextInput = getDecoratedMarkup(controlNode, false);
-      // nextVNode.instance = controlNode;
-      nextInput.ref = controlNode.markup.ref;
-      lifecycle.push(mountWasabyCallback(controlNode));
-      patch(controlNode.markup, nextInput, parentDOM, {}, false, controlNode, lifecycle, controlNode.environment, controlNode);
-      controlNode.markup = nextInput;
-      controlNode.fullMarkup = controlNode.markup;
-  }
-}
-
 // @ts-ignore
 function patchWasabyTemplateNode(lastVNode, nextVNode, parentDOM, context, isSVG, lifecycle, environment, parentControlNode) {
   // @ts-ignore
@@ -530,31 +479,6 @@ function patchClassComponent(lastVNode, nextVNode, parentDOM, context, isSVG: bo
   instance.$UPD = false;
 }
 
-function applyWasabyState(component) {
-  const lifecycle = [];
-  updateWasabyControl(component, component.control._container[0], lifecycle);
-  if (lifecycle.length > 0) {
-    callAll(lifecycle);
-  }
-}
-
-// @ts-ignore
-function queueWasabyControlChanges(controlNode) {
-  if (QUEUE.length === 0) {
-      applyWasabyState(controlNode);
-      return;
-  }
-  // @ts-ignore
-  if (QUEUE.push(controlNode) === 1) {
-      nextTickWasaby(rerenderWasaby);
-  }
-}
-function rerenderWasaby() {
-  let component;
-  while ((component = QUEUE.pop())) {
-    applyWasabyState(component);
-  }
-}
 
 // @ts-ignore
 function patchWasabyControl(lastVNode, nextVNode, parentDOM, context, isSVG, lifecycle, environment, parentControlNode, parentVNode) {
